Intento de commit 2

+++ src/app/src/main/java/com/inventory/User.java
@@ -1,12 +1,11 @@
 package com.inventory;
 
-public class User {
-
-    int id;
+public class User{
+    long id;
     String name;
     String pass;
 
-    public User(int id, String name, String pass) {
+    public User(long id, String name, String pass) {
         this.id = id;
         this.name = name;
         this.pass = pass;
